Moved assertDeliveryError to Delivery reducer
